simplify query condition
